cli dissociation from main into subgroups
