+++ crypy/config.py
@@ -1,4 +1,3 @@
-# -*- coding: utf8 -*-
 import logging
 import os
